Fix #38: moved CrystalSlice to elements with Element base
